timeval needs casts, this doesn't compile on mac+clang without it

+++ src/logging.cpp
@@ -261,7 +261,7 @@ string logdest::format_time(const struct timeval& when) {
             buf, LOGGING_TIMESTR_BUFLEN,
             "%02d:%02d:%02d.%06ld",
             timevals->tm_hour, timevals->tm_min,
-            timevals->tm_sec, when.tv_usec);
+            timevals->tm_sec, (long)when.tv_usec);
 
     if (result >= LOGGING_TIMESTR_BUFLEN) {
         // FIXME this ignores errno
